+++ LanguageLearningAPI/api/models/reading_review.py
@@ -7,7 +7,7 @@ class ReadingReview(models.Model):
     user = models.ForeignKey(User, on_delete = models.CASCADE, related_name = 'rr_users')
     comment = models.TextField()
     publish_date = models.DateTimeField(auto_now_add = True)
-    user_rate = models.FloatField()
+    user_rate = models.IntegerField()
 
     def __str__(self):
         return self.comment
